librmnetctl: introduce and use rta_put() helper
Every single time we add an RTA to a netlink message in the current code,
we copy-paste the same code. This is very messy and not very maintainable.
This patch adds a function to handle adding the RTA to the netlink
message, and updates the code to call this function in each case.

Change-Id: I4276b31ab8e6c61da299c6d2edf15a38c1370d37
Signed-off-by: Sean Tranchetti <stranche@codeaurora.org>